requirements.txt
torch>=1.9.0
pybind11>=2.6.0
numpy>=1.19.0
cupy-cuda12x>=13.0.0 